update to be float instead of int32 on most of subsystem

+++ modules/control_system_manager/control_system_manager.c
@@ -8,10 +8,10 @@
  * - if pointer is null, return the IN value
  * - else apply filter
  */
-static inline uint32_t
-safe_filter(int32_t (*f)(void *, int32_t), void * param, int32_t value)
+static inline float
+safe_filter(float (*f)(void *, float), void * param, float value)
 {
-	int32_t (*f_tmp)(void *, int32_t);
+	float (*f_tmp)(void *, float);
 	void * param_tmp;
 	f_tmp = f;
 	param_tmp = param;
@@ -28,10 +28,10 @@ safe_filter(int32_t (*f)(void *, int32_t), void * param, int32_t value)
  * - if pointer is null, return 0
  * - else return the value processed by the function
  */
-static inline uint32_t
-safe_getprocessout(int32_t (*f)(void *), void * param)
+static inline float
+safe_getprocessout(float (*f)(void *), void * param)
 {
-	int32_t (*f_tmp)(void *);
+	float (*f_tmp)(void *);
 	void * param_tmp;
     f_tmp = f;
 	param_tmp = param;
@@ -49,9 +49,9 @@ safe_getprocessout(int32_t (*f)(void *), void * param)
  * - else call the processin with the parameters
  */
 static inline void
-safe_setprocessin(void (*f)(void *, int32_t), void * param, int32_t value)
+safe_setprocessin(void (*f)(void *, float), void * param, float value)
 {
-	void (*f_tmp)(void *, int32_t);
+	void (*f_tmp)(void *, float);
 	void * param_tmp;
     f_tmp = f;
 	param_tmp = param;
@@ -69,7 +69,7 @@ void cs_init(struct cs* cs)
 }
 
 
-void cs_set_consign_filter(struct cs* cs, int32_t (*consign_filter)(void*, int32_t), void* consign_filter_params)
+void cs_set_consign_filter(struct cs* cs, float (*consign_filter)(void*, float), void* consign_filter_params)
 {
     cs->consign_filter = consign_filter;
     cs->consign_filter_params = consign_filter_params;
@@ -77,27 +77,27 @@ void cs_set_consign_filter(struct cs* cs, int32_t (*consign_filter)(void*, int32
 
 
 
-void cs_set_correct_filter(struct cs* cs, int32_t (*correct_filter)(void*, int32_t), void* correct_filter_params)
+void cs_set_correct_filter(struct cs* cs, float (*correct_filter)(void*, float), void* correct_filter_params)
 {
     cs->correct_filter = correct_filter;
     cs->correct_filter_params = correct_filter_params;
 }
 
 
-void cs_set_feedback_filter(struct cs* cs, int32_t (*feedback_filter)(void*, int32_t), void* feedback_filter_params)
+void cs_set_feedback_filter(struct cs* cs, float (*feedback_filter)(void*, float), void* feedback_filter_params)
 {
     cs->feedback_filter = feedback_filter;
     cs->feedback_filter_params = feedback_filter_params;
 }
 
 /** Set the cs output_filter fields in the cs structure */
-void  cs_set_output_filter(struct cs* cs, int32_t (*output_filter)(void*, int32_t), void* output_filter_params) {
+void  cs_set_output_filter(struct cs* cs, float (*output_filter)(void*, float), void* output_filter_params) {
     cs->output_filter = output_filter;
     cs->output_filter_params = output_filter_params;
 }
 
 
-void cs_set_process_in(struct cs* cs, void (*process_in)(void*, int32_t), void* process_in_params)
+void cs_set_process_in(struct cs* cs, void (*process_in)(void*, float), void* process_in_params)
 {
     cs->process_in = process_in;
     cs->process_in_params = process_in_params;
@@ -105,7 +105,7 @@ void cs_set_process_in(struct cs* cs, void (*process_in)(void*, int32_t), void* 
 
 
 
-void cs_set_process_out(struct cs* cs, int32_t (*process_out)(void*), void* process_out_params)
+void cs_set_process_out(struct cs* cs, float (*process_out)(void*), void* process_out_params)
 {
     cs->process_out = process_out;
     cs->process_out_params = process_out_params;
@@ -113,9 +113,9 @@ void cs_set_process_out(struct cs* cs, int32_t (*process_out)(void*), void* proc
 
 
 
-int32_t cs_do_process(struct cs* cs, int32_t consign)
+float cs_do_process(struct cs* cs, float consign)
 {
-    int32_t process_out_value = 0;
+    float process_out_value = 0;
 
     if (cs->enabled) {
         cs->consign_value = consign;
@@ -152,39 +152,39 @@ void cs_manage(void * data)
 
 
 
-int32_t cs_get_out(struct cs* cs)
+float cs_get_out(struct cs* cs)
 {
    return cs->out_value;
 }
 
-int32_t cs_get_error(struct cs* cs)
+float cs_get_error(struct cs* cs)
 {
     return cs->error_value;
 }
 
 
 
-int32_t cs_get_consign(struct cs* cs)
+float cs_get_consign(struct cs* cs)
 {
     return cs->consign_value;
 }
 
-int32_t cs_get_filtered_consign(struct cs* cs)
+float cs_get_filtered_consign(struct cs* cs)
 {
     return cs->filtered_consign_value;
 }
 
-int32_t cs_get_filtered_feedback(struct cs* cs)
+float cs_get_filtered_feedback(struct cs* cs)
 {
     return cs->filtered_feedback_value;
 }
 
-int32_t cs_get_feedback(struct cs* cs)
+float cs_get_feedback(struct cs* cs)
 {
 	return safe_getprocessout(cs->process_out, cs->process_out_params);
 }
 
-void cs_set_consign(struct cs* cs, int32_t v)
+void cs_set_consign(struct cs* cs, float v)
 {
     cs->consign_value = v;
 }
